Use Glossary in Json
