--- webui/react/src/components/Json.module.scss
@@ -1,45 +1,0 @@
-.base {
-  margin: 0;
-  padding-inline-start: 0;
-}
-.item {
-  align-items: center;
-  border-color: var(--theme-stage-border);
-  border-style: dashed;
-  border-width: 0 0 1px;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  list-style-type: none;
-  margin-bottom: 2px;
-  padding-bottom: 2px;
-  width: 100%;
-
-  span {
-    display: block;
-  }
-}
-.item:last-child {
-  border: 0;
-  margin: 0;
-  padding: 0;
-}
-.label {
-  color: var(--theme-stage-on);
-  margin-right: 8px;
-}
-.value {
-  overflow-wrap: break-word;
-  text-align: right;
-  word-break: break-all;
-}
-.hideDivider {
-  .item {
-    border-width: 0;
-  }
-}
-.alternateBackground {
-  li:nth-child(even) {
-    background-color: var(--theme-colors-monochrome-13);
-  }
-}
